refactored HEP units into separate files and corrected namespace where necessary
