+++ src/systems/si/include/units/isq/si/mass.h
@@ -85,25 +85,6 @@ struct dim_mass : isq::dim_mass<kilogram> {};
 template<UnitOf<dim_mass> U, Representation Rep = double>
 using mass = quantity<dim_mass, U, Rep>;
 
-namespace hep {
-struct eV_per_c2 : named_scaled_unit<eV_per_c2, "eV/c^2", prefix, ratio(1'7826'619'216'279, 1'000'000'000'000, -35), kilogram> {};
-struct feV_per_c2 : prefixed_unit<feV_per_c2, femto, eV_per_c2> {};
-struct peV_per_c2 : prefixed_unit<peV_per_c2, pico, eV_per_c2> {};
-struct neV_per_c2 : prefixed_unit<neV_per_c2, nano, eV_per_c2> {};
-struct ueV_per_c2 : prefixed_unit<ueV_per_c2, micro, eV_per_c2> {};
-struct meV_per_c2 : prefixed_unit<meV_per_c2, milli, eV_per_c2> {};  // approximate mass of an electron/positron (0.511 MeV/c2)
-struct keV_per_c2 : prefixed_unit<keV_per_c2, kilo, eV_per_c2> {};
-struct MeV_per_c2 : prefixed_unit<MeV_per_c2, mega, eV_per_c2> {};
-struct GeV_per_c2 : prefixed_unit<GeV_per_c2, giga, eV_per_c2> {};  // approximate mass of a proton (0.938 GeV/c2) or neutron
-struct TeV_per_c2 : prefixed_unit<TeV_per_c2, tera, eV_per_c2> {};
-struct PeV_per_c2 : prefixed_unit<PeV_per_c2, peta, eV_per_c2> {};
-struct EeV_per_c2 : prefixed_unit<EeV_per_c2, exa, eV_per_c2> {};
-struct YeV_per_c2 : prefixed_unit<YeV_per_c2, yotta, eV_per_c2> {};
-struct electron_mass : named_scaled_unit<eV_per_c2, "m_e", prefix, ratio(9'109'383'701'528, 1'000'000'000'000, -31), kilogram> {};
-struct proton_mass : named_scaled_unit<eV_per_c2, "m_p", prefix, ratio(1'672'621'923'695, 1'000'000'000'000, -27), kilogram> {};
-struct neutron_mass : named_scaled_unit<eV_per_c2, "m_n", prefix, ratio(1'674'927'498'049, 1'000'000'000'000, -27), kilogram> {};
-}  // namespace hep
-
 #ifndef UNITS_NO_LITERALS
 
 inline namespace literals {
@@ -281,51 +262,7 @@ constexpr auto operator"" _q_Yt(long double l) { return mass<yottatonne, long do
 constexpr auto operator"" _q_Da(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<dalton, std::int64_t>(static_cast<std::int64_t>(l)); }
 constexpr auto operator"" _q_Da(long double l) { return mass<dalton, long double>(l); }
 
-namespace units::isq::inline si::inline hep::literals {
-
-}
-
 }  // namespace literals
-
-namespace hep::literals {
-constexpr auto operator"" _q_feV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<feV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_feV_per_c2(long double l) { return mass<feV_per_c2, long double>(l); }
-constexpr auto operator"" _q_peV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<peV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_peV_per_c2(long double l) { return mass<peV_per_c2, long double>(l); }
-constexpr auto operator"" _q_neV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<neV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_neV_per_c2(long double l) { return mass<neV_per_c2, long double>(l); }
-constexpr auto operator"" _q_ueV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<ueV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_ueV_per_c2(long double l) { return mass<ueV_per_c2, long double>(l); }
-constexpr auto operator"" _q_meV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<meV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_meV_per_c2(long double l) { return mass<meV_per_c2, long double>(l); }
-
-// eV_per_c2
-constexpr auto operator"" _q_eV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<eV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_eV_per_c2(long double l) { return mass<eV_per_c2, long double>(l); }
-
-constexpr auto operator"" _q_keV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<keV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_keV_per_c2(long double l) { return mass<keV_per_c2, long double>(l); }
-constexpr auto operator"" _q_MeV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<MeV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_MeV_per_c2(long double l) { return mass<MeV_per_c2, long double>(l); }
-constexpr auto operator"" _q_GeV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<GeV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_GeV_per_c2(long double l) { return mass<GeV_per_c2, long double>(l); }
-constexpr auto operator"" _q_TeV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<TeV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_TeV_per_c2(long double l) { return mass<TeV_per_c2, long double>(l); }
-constexpr auto operator"" _q_PeV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<PeV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_PeV_per_c2(long double l) { return mass<PeV_per_c2, long double>(l); }
-constexpr auto operator"" _q_EeV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<EeV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_EeV_per_c2(long double l) { return mass<EeV_per_c2, long double>(l); }
-constexpr auto operator"" _q_YeV_per_c2(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<YeV_per_c2, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_YeV_per_c2(long double l) { return mass<YeV_per_c2, long double>(l); }
-
-// special HEP masses
-constexpr auto operator"" _q_electron_mass(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<electron_mass, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_electron_mass(long double l) { return mass<electron_mass, long double>(l); }
-constexpr auto operator"" _q_proton_mass(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<proton_mass, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_proton_mass(long double l) { return mass<proton_mass, long double>(l); }
-constexpr auto operator"" _q_neutron_mass(unsigned long long l) { gsl_ExpectsAudit(std::in_range<std::int64_t>(l)); return mass<neutron_mass, std::int64_t>(static_cast<std::int64_t>(l)); }
-constexpr auto operator"" _q_neutron_mass(long double l) { return mass<neutron_mass, long double>(l); }
-} // namespace hep::literals
 
 #endif // UNITS_NO_LITERALS
